requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Async PostgreSQL
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# Settings management
pydantic-settings==2.7.1

# Migrations
alembic==1.14.1

# PostgreSQL sync driver (used by alembic)
psycopg2-binary==2.9.10

# Async HTTP client (API runner)
httpx==0.28.1

# AI / LLM
openai==1.82.0

# Scheduler
apscheduler==3.11.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
email-validator==2.2.0

# Rate limiting
slowapi==0.1.9
